+++ e2e/snapshot-serializers-esm/plugins/foo/index.mjs
@@ -6,5 +6,6 @@
  *
  */
 
-import {createPlugin} from '../../utils';
+import {createPlugin} from '../../utils.mjs';
+
 export default createPlugin('foo');